+++ frontend/src/screens/signup.tsx
@@ -1,14 +1,7 @@
 /* eslint-disable react-native/no-unused-styles */
 import * as React from 'react';
 import { useState } from 'react';
-import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-  TextInput,
-  Alert,
-} from 'react-native';
+import { StyleSheet, Text, View, TouchableOpacity, TextInput, Alert } from 'react-native';
 import { useFonts } from 'expo-font';
 import { type RootStackParamList } from '../components/types';
 import { type StackNavigationProp } from '@react-navigation/stack';
@@ -20,7 +13,7 @@ import ObjectID from 'bson-objectid';
 
 export type StackNavigation = StackNavigationProp<RootStackParamList>;
 
-export default function SignUp(): JSX.Element{
+export default function SignUp(): JSX.Element {
   const [fullName, setFullName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
@@ -37,7 +30,7 @@ export default function SignUp(): JSX.Element{
     } else if (password !== verifiedPassword) {
       Alert.alert('Please correctly rewrite your Password');
       return false;
-    }else {
+    } else {
       return true;
     }
   };
@@ -47,20 +40,18 @@ export default function SignUp(): JSX.Element{
       try {
         await firebaseService.createUser(email, password);
         await createUser({
-          _id: new ObjectID,
+          _id: new ObjectID(),
           full_name: fullName,
           email,
           badges: [],
           friends: [],
           score: 0,
         });
-        navigation.navigate('DashBoard');
+        navigation.navigate('SignUpQuestions');
       } catch (error) {
         console.error('Error when creating User:', error);
       }
-
     }
-
   };
 
   const [loaded] = useFonts({
@@ -74,52 +65,61 @@ export default function SignUp(): JSX.Element{
 
   return (
     <View style={styles.headerContainer}>
-    <View style={styles.headerBox}>
-      <Text style={styles.header}> Sign Up </Text>
+      <View style={styles.headerBox}>
+        <Text style={styles.header}> Sign Up </Text>
 
-      <View style={styles.textbox}>
-        <TextInput
+        <View style={styles.textbox}>
+          <TextInput
             style={styles.textInputBox}
             placeholder="Full Name"
             value={fullName}
-            onChangeText={(fullName) => {setFullName(fullName)}}
+            onChangeText={(fullName) => {
+              setFullName(fullName);
+            }}
           />
-      </View>
+        </View>
 
-      <View style={styles.textbox}>
-      <TextInput
-        style={styles.textInputBox}
-        placeholder="Email"
-        value={email}
-        onChangeText={(email) => {setEmail(email)}}
-      />
-      </View>
-      <View style={styles.textbox}>
-      <TextInput
-        style={styles.textInputBox}
-        placeholder="Password"
-        value={password}
-        onChangeText={(password) => {setPassword(password)}}
-      />
-      </View>
-      <View style={styles.textbox}>
-      <TextInput 
-        style={styles.textInputBox} 
-        placeholder="Re-enter Password"
-        value={verifiedPassword}
-        onChangeText={(verifiedPassword) => {setVerifiedPassword(verifiedPassword)}} />
-      </View>
+        <View style={styles.textbox}>
+          <TextInput
+            style={styles.textInputBox}
+            placeholder="Email"
+            value={email}
+            onChangeText={(email) => {
+              setEmail(email);
+            }}
+          />
+        </View>
+        <View style={styles.textbox}>
+          <TextInput
+            style={styles.textInputBox}
+            placeholder="Password"
+            value={password}
+            onChangeText={(password) => {
+              setPassword(password);
+            }}
+          />
+        </View>
+        <View style={styles.textbox}>
+          <TextInput
+            style={styles.textInputBox}
+            placeholder="Re-enter Password"
+            value={verifiedPassword}
+            onChangeText={(verifiedPassword) => {
+              setVerifiedPassword(verifiedPassword);
+            }}
+          />
+        </View>
 
-      <TouchableOpacity
+        <TouchableOpacity
           style={styles.buttoning}
-          onPress={ () => {void handleSignUp();
+          onPress={() => {
+            void handleSignUp();
           }}
-        
         >
           <Text style={styles.altContainerText}> Next </Text>
         </TouchableOpacity>
+      </View>
     </View>
-  </View>   
   );
 }
 
@@ -127,38 +127,38 @@ const styles = StyleSheet.create({
   altContainer: {
     alignItems: 'center',
     backgroundColor: Colors.LIGHTFGREEN,
-    flexDirection: 'row', 
+    flexDirection: 'row',
     marginBottom: 30,
   },
-    
-  altContainerFloater:{
+
+  altContainerFloater: {
     backgroundColor: Colors.DARKLIMEGREEN,
-    flex: 1, 
-    height: 1, 
+    flex: 1,
+    height: 1,
   },
   altContainerSub: {
     backgroundColor: Colors.DARKLIMEGREEN,
-    flex: 1, 
-    height: 1, 
+    flex: 1,
+    height: 1,
   },
   altContainerText: {
     color: Colors.DARKLIMEGREEN,
     fontFamily: 'Montserrat',
     fontSize: 18,
     marginHorizontal: 5,
-    textAlign: 'center',  
+    textAlign: 'center',
   },
-  buttoning:{
-    backgroundColor: Colors.DARKGREEN, 
+  buttoning: {
+    backgroundColor: Colors.DARKGREEN,
     borderRadius: 10,
     marginBottom: 20,
-    padding: 18,  
-  }, 
+    padding: 18,
+  },
   buttoningText: {
-    color: Colors.WHITE, 
-    fontSize: 16, 
-    fontWeight: '700', 
-    textAlign: 'center', 
+    color: Colors.WHITE,
+    fontSize: 16,
+    fontWeight: '700',
+    textAlign: 'center',
   },
   footer: {
     color: Colors.DARKLIMEGREEN,
@@ -176,26 +176,26 @@ const styles = StyleSheet.create({
     textAlign: 'center',
   },
   footerContainer: {
-    flexDirection: 'row', 
-    justifyContent: 'center'
+    flexDirection: 'row',
+    justifyContent: 'center',
   },
-  googleIcon:{
-    alignItems: 'center', 
-    justifyContent: 'center'
+  googleIcon: {
+    alignItems: 'center',
+    justifyContent: 'center',
   },
   header: {
-    color: Colors.DARKGREEN, 
-    fontFamily: 'Montserrat', 
-    fontSize: 30, 
+    color: Colors.DARKGREEN,
+    fontFamily: 'Montserrat',
+    fontSize: 30,
     fontWeight: '700',
     marginBottom: 30,
   },
   headerBox: {
-    paddingHorizontal: 30
+    paddingHorizontal: 30,
   },
-  headerContainer:{
+  headerContainer: {
     flex: 1,
-    justifyContent: 'center'
+    justifyContent: 'center',
   },
   icon: {
     height: 24,
@@ -203,8 +203,8 @@ const styles = StyleSheet.create({
     width: 24,
   },
   textInputBox: {
-    flex:1, 
-    paddingVertical:0,
+    flex: 1,
+    paddingVertical: 0,
   },
   textbox: {
     borderBottomColor: Colors.GREY,
@@ -214,4 +214,3 @@ const styles = StyleSheet.create({
     paddingBotton: 8,
   },
 });
-
