+++ frontend/src/screens/transportationForum.tsx
@@ -1,23 +1,11 @@
-import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-  ScrollView,
-  TextInput,
-  Modal,
-  Linking,
-} from 'react-native';
-import { CheckBox } from 'react-native-elements';
+import { StyleSheet, Text, View, TouchableOpacity, ScrollView } from 'react-native';
 import * as React from 'react';
-import Icon from 'react-native-vector-icons/FontAwesome';
 import { useState } from 'react';
 import { type StackNavigationProp } from '@react-navigation/stack';
 import { type RootStackParamList } from '../components/types';
 import { useNavigation } from '@react-navigation/native';
 import { useFonts } from 'expo-font';
 import Colors from '../../assets/colorConstants';
-import transportationQuestions from '../../assets/questions';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import Slider from '@react-native-community/slider';
 
@@ -66,23 +54,6 @@ export default function TransportationForum(): JSX.Element {
     }
   };
 
-  const openLink = async (url: string): Promise<void> => {
-    const supported = await Linking.canOpenURL(url);
-
-    if (supported) {
-      await Linking.openURL(url);
-    } else {
-      console.error('Cannot open the URL');
-    }
-  };
-
-  const handleLinkPress = (): void => {
-    const url = 'https://fcr-ccc.nrcan-rncan.gc.ca/en';
-    openLink(url).catch((error) => {
-      console.error('Error opening link:', error);
-    });
-  };
-
   const [loaded] = useFonts({
     Montserrat: require('../../assets/fonts/MontserratThinRegular.ttf'),
     Josefin: require('../../assets/fonts/JosefinSansThinRegular.ttf'),
@@ -90,24 +61,14 @@ export default function TransportationForum(): JSX.Element {
 
   const navigation = useNavigation<StackNavigation>();
 
-  const data = transportationQuestions;
-
-  const [responses, setResponses] = useState<string[]>(new Array(data.length).fill(''));
-
-  const [fuelType, setFuelType] = useState<string>('');
-  const [fuelEfficiency, setFuelEfficiency] = useState(0);
   const [carTravel, setCarTravel] = useState(0);
   const [busTravel, setBusTravel] = useState(0);
   const [trainTravel, setTrainTravel] = useState(0);
   const [bikeTravel, setBikeTravel] = useState(0);
 
-  const [modalVisible, setModalVisible] = useState(false);
-
   const handleSurveySubmit = (): void => {
     // Process survey responses, e.g., send them to a server
     console.log('Survey Responses:', {
-      fuelType,
-      fuelEfficiency,
       carTravel,
       busTravel,
       trainTravel,
@@ -115,12 +76,6 @@ export default function TransportationForum(): JSX.Element {
     });
 
     navigation.navigate('FoodForum');
-  };
-
-  const handleOptionSelect = (questionId: number, optionIndex: number): void => {
-    const updatedResponses = [...responses];
-    updatedResponses[questionId] = data[questionId].options[optionIndex];
-    setResponses(updatedResponses);
   };
 
   if (!loaded) {
@@ -131,72 +86,6 @@ export default function TransportationForum(): JSX.Element {
     <SafeAreaView style={styles.container}>
       <ScrollView style={styles.scrollContainer}>
         <Text style={styles.header}>Calculate your emissions from transportation:</Text>
-
-        <View style={styles.questionContainer}>
-          <Text style={styles.question}>{data[0].question}</Text>
-          {data[0].options.map((option, index) => (
-            <CheckBox
-              checkedColor={Colors.DARKGREEN}
-              textStyle={styles.answer}
-              containerStyle={
-                responses[data[0].id] === option ? styles.selectedOption : styles.unSelectedOption
-              }
-              key={index}
-              title={option}
-              checked={responses[data[0].id] === option}
-              onPress={() => {
-                handleOptionSelect(data[0].id, index);
-                setFuelType(data[0].options[index]);
-              }}
-            />
-          ))}
-        </View>
-
-        <View style={styles.questionContainer}>
-          <View style={styles.questionWithIcon}>
-            <Text style={styles.question}>
-              Please enter your vehicle&apos;s fuel efficiency. If you don&apos;t have a vehicle,
-              enter 0.
-            </Text>
-            <TouchableOpacity
-              style={styles.questionIcon}
-              onPress={() => setModalVisible(!modalVisible)}
-            >
-              <Icon name="question-circle" size={30} color={Colors.DARKGREEN} />
-            </TouchableOpacity>
-          </View>
-          <TextInput
-            style={styles.input}
-            keyboardType="numeric"
-            placeholder="Input"
-            onChangeText={(text) => {
-              setFuelEfficiency(Number(text));
-            }}
-          />
-        </View>
-
-        <Modal transparent={true} visible={modalVisible}>
-          <View style={styles.modalBackground}>
-            <View style={styles.modalContainer}>
-              <Text style={styles.infoText}>
-                If you don&apos;t know your vehicle&apos;s fuel efficiency, it&apos;s available
-                online{' '}
-                <Text style={styles.linkText} onPress={handleLinkPress}>
-                  here
-                </Text>
-                . Select the &quot;combination&quot; value under Comsumption in L/100km. The average
-                fuel consumption of non-plug-in hybrid personal vehicles in Canada is 8.9 L / 100
-                km.
-              </Text>
-              <TouchableOpacity
-                style={styles.closeIcon}
-                onPress={() => setModalVisible(!modalVisible)}
-              >
-                <Icon name="times-circle" size={30} color={Colors.DARKGREEN} />
-              </TouchableOpacity>
-            </View>
-          </View>
-        </Modal>
 
         <Text style={styles.header}>
           On average, how much distance do you travel using the following methods per week:
@@ -246,19 +135,6 @@ const styles = StyleSheet.create({
     width: '100%',
     paddingHorizontal: 12,
   },
-  modalBackground: {
-    backgroundColor: Colors.BLACKTRANS,
-    flex: 1,
-  },
-  modalContainer: {
-    backgroundColor: Colors.WHITE,
-    marginHorizontal: 50,
-    marginVertical: 180,
-    padding: 20,
-    borderRadius: 10,
-    flex: 1,
-    flexDirection: 'row',
-  },
   scrollContainer: {
     paddingTop: 30,
     flex: 1,
@@ -282,20 +158,6 @@ const styles = StyleSheet.create({
     color: Colors.DARKGREEN,
     marginBottom: 20,
   },
-  answer: {
-    fontFamily: 'Montserrat',
-    fontSize: 17,
-    fontWeight: '700',
-    color: Colors.DARKGREEN,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: Colors.GREY,
-    padding: 8,
-    margin: 10,
-    width: 200,
-    backgroundColor: Colors.WHITE,
-  },
   buttoning: {
     backgroundColor: Colors.DARKGREEN,
     borderRadius: 10,
@@ -308,31 +170,8 @@ const styles = StyleSheet.create({
     fontWeight: '700',
     textAlign: 'center',
   },
-  selectedOption: {
-    backgroundColor: Colors.FILLGREEN,
-  },
-  unSelectedOption: {
-    backgroundColor: Colors.LIGHTFGREEN,
-  },
   questionContainer: {
     paddingBottom: 30,
-  },
-  questionWithIcon: {
-    flexDirection: 'row',
-  },
-  infoText: {
-    fontSize: 20,
-    color: Colors.DARKGREEN,
-  },
-  linkText: {
-    color: Colors.BLUE,
-  },
-  closeIcon: {
-    marginLeft: 'auto',
-  },
-  questionIcon: {
-    marginLeft: 15,
-    paddingTop: 5,
   },
   silder: {
     height: 50,
